Presantation of the major project
I have in this presantation answar the following points:

A description of how you want to present the project to the world
A reflection of the experiences you had while creating the project
A list of what could be improved in the project if you had more time?
A conclusion on the budget. Did you manage to stick to the budget? Why? / Why not?


( This is the same presantation from 19 May)